app/requirements.txt
Django==4.1.7
pymongo==3.13.0
sqlparse==0.2.4
djangorestframework==3.14.0
djongo==1.3.6
django-allauth
dj-rest-auth
djangorestframework-simplejwt
django-cors-headers